refactor(blocker): Change blocklist source files to block_manager
