+++ src/blocker/Makefile.am
@@ -10,8 +10,8 @@ sshg_blocker_SOURCES = \
 	../common/heap.h \
 	attack.c \
 	blocker.c \
-	blocklist.c \
-	blocklist.h \
+	block_manager.c \
+	block_manager.h \
 	fnv.h \
 	hash_32a.c \
 	sshguard_blacklist.c \
