+++ src/blocker/block_manager.h
